Split out linksight.api.views
