+++ linksight/api/views/__init__.py
@@ -1,0 +1,1 @@
+from linksight.api.views import dataset, match, user
